로컬캐시 추가 및 querydsl 변경

+++ src/main/kotlin/com/test/searchAPI/application/keyword/model/PopularKeywordResponse.kt
@@ -1,11 +1,11 @@
 package com.test.searchAPI.application.keyword.model
 
-import com.test.searchAPI.domain.keyword.entity.PopularKeyword
+import com.test.searchAPI.domain.keyword.model.PopularKeyword
 
 data class PopularKeywordResponse (
     val index: Int,
     val keyword: String,
-    val searchCount: Int,
+    val searchCount: Long,
 ) {
     companion object {
         fun of(index: Int, keyword: PopularKeyword) = PopularKeywordResponse(index, keyword.keyword, keyword.searchCount)
